make little changes for rendering fire

+++ src/edu/neu/mhealth/debug/opengl/OpenGLFire.java
@@ -5,15 +5,15 @@ public class OpenGLFire {
 	 * Because the units of width are different between OpenCV and OpenGL, we
 	 * record ratio for the position according to the screen
 	 */
-	public float ratioX;
+	public double ratioX;
 
 	/**
 	 * Because the units of height are different between OpenCV and OpenGL, we
 	 * record ratio for the position according to the screen
 	 */
-	public float ratioY;
+	public double ratioY;
 
-	public OpenGLFire(float ratioX, float ratioY) {
+	public OpenGLFire(double ratioX, double ratioY) {
 		this.ratioX = ratioX;
 		this.ratioY = ratioY;
 	}
